requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
osmnx
openai
requests
Pillow
